Suggestions applied
