+++ pennylane/devices/default_qubit.py
@@ -605,7 +605,7 @@ class DefaultQubit(QubitDevice):
         """
         wires = wires.tolist()
         state = self._flatten(self._pre_rotated_state)
-        return qml.math.to_density_matrix(state, wires=wires)
+        return qml.math.to_density_matrix(state, indices=wires)
 
     def _apply_state_vector(self, state, device_wires):
         """Initialize the internal state vector in a specified state.
