+++ tests/math/test_functions.py
@@ -2261,183 +2261,3 @@ class TestSortFunction:
         result = fn.sort(input)
 
         assert all(result == test_output)
-
-
-ones_functions = [onp.ones, np.ones, jnp.ones, torch.ones, tf.ones]
-
-state_00 = [1, 0, 0, 0]
-state_01 = [0, 1, 0, 0]
-state_10 = [0, 0, 1, 0]
-state_11 = [0, 0, 0, 1]
-
-state_00_10 = [1, 0, 1, 0] / onp.sqrt(2)
-state_01_11 = [0, 1, 0, 1] / onp.sqrt(2)
-
-mat_00 = onp.zeros((4, 4))
-mat_00[0, 0] = 1
-
-mat_01 = onp.zeros((4, 4))
-mat_01[1, 1] = 1
-
-mat_10 = onp.zeros((4, 4))
-mat_10[2, 2] = 1
-
-mat_11 = onp.zeros((4, 4))
-mat_11[3, 3] = 1
-
-mat_0 = onp.zeros((2, 2))
-mat_0[0, 0] = 1
-
-mat_1 = onp.zeros((2, 2))
-mat_1[1, 1] = 1
-
-mat_00_10 = onp.zeros((4, 4))
-mat_00_10[0, 0] = 0.5
-mat_00_10[2, 2] = 0.5
-mat_00_10[0, 2] = 0.5
-mat_00_10[2, 0] = 0.5
-
-mat_01_11 = onp.zeros((4, 4))
-mat_01_11[1, 1] = 0.5
-mat_01_11[3, 3] = 0.5
-mat_01_11[1, 3] = 0.5
-mat_01_11[3, 1] = 0.5
-
-mat_0_1 = [[0.5, 0.5], [0.5, 0.5]]
-
-# fmt: off
-state_vectors = [
-    (state_00, (mat_0, mat_0, mat_00)),
-    (state_01, (mat_0, mat_1, mat_01)),
-    (state_10, (mat_1, mat_0, mat_10)),
-    (state_11, (mat_1, mat_1, mat_11)),
-    (onp.array(state_00), (mat_0, mat_0, mat_00)),
-    (onp.array(state_01), (mat_0, mat_1, mat_01)),
-    (onp.array(state_10), (mat_1, mat_0, mat_10)),
-    (onp.array(state_11), (mat_1, mat_1, mat_11)),
-    (np.array(state_00), (mat_0, mat_0, mat_00)),
-    (np.array(state_01), (mat_0, mat_1, mat_01)),
-    (np.array(state_10), (mat_1, mat_0, mat_10)),
-    (np.array(state_11), (mat_1, mat_1, mat_11)),
-    (jnp.array(state_00), (mat_0, mat_0, mat_00)),
-    (jnp.array(state_01), (mat_0, mat_1, mat_01)),
-    (jnp.array(state_10), (mat_1, mat_0, mat_10)),
-    (jnp.array(state_11), (mat_1, mat_1, mat_11)),
-    (torch.tensor(state_00), (mat_0, mat_0, mat_00)),
-    (torch.tensor(state_01), (mat_0, mat_1, mat_01)),
-    (torch.tensor(state_10), (mat_1, mat_0, mat_10)),
-    (torch.tensor(state_11), (mat_1, mat_1, mat_11)),
-    (tf.Variable(state_00), (mat_0, mat_0, mat_00)),
-    (tf.Variable(state_01), (mat_0, mat_1, mat_01)),
-    (tf.Variable(state_10), (mat_1, mat_0, mat_10)),
-    (tf.Variable(state_11), (mat_1, mat_1, mat_11)),
-    (tf.constant(state_00), (mat_0, mat_0, mat_00)),
-    (tf.constant(state_01), (mat_0, mat_1, mat_01)),
-    (tf.constant(state_10), (mat_1, mat_0, mat_10)),
-    (tf.constant(state_11), (mat_1, mat_1, mat_11)),
-    (state_00_10, (mat_0_1, mat_0, mat_00_10)),
-    (state_01_11, (mat_0_1, mat_1, mat_01_11)),
-]
-
-single_wires_list = [
-    [0],
-    [1],
-]
-
-multiple_wires_list = [
-    [0, 1]
-]
-# fmt: on
-
-
-class TestDensityMatrixFromStateVectors:
-    """Tests for creating a density matrix from state vectors."""
-
-    @pytest.mark.parametrize("state_vector, expected_density_matrix", state_vectors)
-    @pytest.mark.parametrize("wires", single_wires_list)
-    def test_density_matrix_from_state_vector_single_wires(
-        self, state_vector, wires, expected_density_matrix
-    ):
-        """Test the density matrix from state vectors for single wires."""
-        density_matrix = fn.quantum._density_matrix_from_state_vector(state_vector, wires=wires)
-        assert np.allclose(density_matrix, expected_density_matrix[wires[0]])
-
-    @pytest.mark.parametrize("state_vector, expected_density_matrix", state_vectors)
-    @pytest.mark.parametrize("wires", multiple_wires_list)
-    def test_density_matrix_from_state_vector_full_wires(
-        self, state_vector, wires, expected_density_matrix
-    ):
-        """Test the density matrix from state vectors for full wires."""
-        density_matrix = fn.quantum._density_matrix_from_state_vector(state_vector, wires=wires)
-        assert np.allclose(density_matrix, expected_density_matrix[2])
-
-    @pytest.mark.parametrize("state_vector, expected_density_matrix", state_vectors)
-    @pytest.mark.parametrize("wires", single_wires_list)
-    def test_density_matrix_from_state_vector_single_wires(
-        self, state_vector, wires, expected_density_matrix
-    ):
-        """Test the density matrix from state vectors for single wires."""
-        density_matrix = fn.to_density_matrix(state_vector, wires=wires)
-        assert np.allclose(density_matrix, expected_density_matrix[wires[0]])
-
-    @pytest.mark.parametrize("state_vector, expected_density_matrix", state_vectors)
-    @pytest.mark.parametrize("wires", multiple_wires_list)
-    def test_density_matrix_from_state_vector_full_wires(
-        self, state_vector, wires, expected_density_matrix
-    ):
-        """Test the density matrix from state vectors for full wires."""
-        density_matrix = fn.to_density_matrix(state_vector, wires=wires)
-        assert np.allclose(density_matrix, expected_density_matrix[2])
-
-    @pytest.mark.parametrize("state_vector, expected_density_matrix", state_vectors)
-    @pytest.mark.parametrize("wires", multiple_wires_list)
-    def test_density_matrix_from_state_vector_check_state(
-        self, state_vector, wires, expected_density_matrix
-    ):
-        """Test the density matrix from state vectors for single wires with state checking"""
-        density_matrix = fn.quantum._density_matrix_from_state_vector(
-            state_vector, wires=wires, check_state=True
-        )
-        assert np.allclose(density_matrix, expected_density_matrix[2])
-
-    def test_state_vector_wrong_shape(self):
-        """Test that wrong shaped state vector raises an error with check_state=True"""
-        state_vector = [1, 0, 0]
-
-        with pytest.raises(ValueError, match="State vector must be"):
-            fn.quantum._density_matrix_from_state_vector(state_vector, wires=[0], check_state=True)
-
-    def test_state_vector_wrong_norm(self):
-        """Test that state vector with wrong norm raises an error with check_state=True"""
-        state_vector = [0.1, 0, 0, 0]
-
-        with pytest.raises(ValueError, match="Sum of amplitudes-squared does not equal one."):
-            fn.quantum._density_matrix_from_state_vector(state_vector, wires=[0], check_state=True)
-
-    def test_density_matrix_from_state_vector_jax_jit(self):
-        """Test jitting the density matrix from state vector function."""
-        from jax import jit
-        import jax.numpy as jnp
-
-        state_vector = jnp.array([1, 0, 0, 0])
-
-        jitted_dens_matrix_func = jit(
-            fn.quantum._density_matrix_from_state_vector, static_argnums=[1, 2]
-        )
-
-        density_matrix = jitted_dens_matrix_func(state_vector, wires=(0, 1), check_state=True)
-        assert np.allclose(density_matrix, [[1, 0, 0, 0], [0, 0, 0, 0], [0, 0, 0, 0], [0, 0, 0, 0]])
-
-    def test_wrong_shape_jax_jit(self):
-        """Test jitting the density matrix from state vector with wrong shape."""
-        from jax import jit
-        import jax.numpy as jnp
-
-        state_vector = jnp.array([1, 0, 0])
-
-        jitted_dens_matrix_func = jit(
-            fn.quantum._density_matrix_from_state_vector, static_argnums=[1, 2]
-        )
-
-        with pytest.raises(ValueError, match="State vector must be"):
-            jitted_dens_matrix_func(state_vector, wires=(0, 1), check_state=True)
